Revert "Bug 2020608 - Part 17: Use std::rot[lr] for Rotate(Left|Right). r=sergesanspaille" for causing bustages at PowerOfTwo.h.
This reverts commit eb884a9f657b9d61dc054e79d33d3186373ba9c9.

Revert "Bug 2020608 - Part 16: Use std::bit_ceil for RoundUpPow2. r=sergesanspaille"

This reverts commit 35a310aa8a2a59d8e01f3a88c39b28582d0b5b9a.

Revert "Bug 2020608 - Part 15: Require unsigned integral types for CeilingLog2 and FindMostSignificantBit. r=sergesanspaille,nbp"

This reverts commit f682901ca2830c6ae8834f65c52fe8388e961653.

Revert "Bug 2020608 - Part 14: Remove mozilla::IsPowerOfTwo. r=sergesanspaille"

This reverts commit b998e4bd39134ae12ecbbfb63d8916f90b2a982c.

Revert "Bug 2020608 - Part 13: Replace IsPowerOfTwo with std::has_single_bit in js/. r=sergesanspaille"

This reverts commit 00db5e1f32b08784aa1e208c1ab1d6ac4729a93e.

Revert "Bug 2020608 - Part 12: Replace IsPowerOfTwo with std::has_single_bit outside of js/. r=sergesanspaille,profiler-reviewers,win-reviewers,gstoll,canaltinova"

This reverts commit cec056f5d6e6bba71a6bf89a85cf2c99b215d91e.

Revert "Bug 2020608 - Part 11: Simplify PowerOfTwo. r=sergesanspaille,profiler-reviewers,canaltinova"

This reverts commit a2ee5b905373d470072f5d75267390239175b9d9.

Revert "Bug 2020608 - Part 10: Prefer std::bit_width in MathAlgorithms.h. r=sergesanspaille"

This reverts commit 89d966bb97945c200fdf088a53294d29431b66b0.

Revert "Bug 2020608 - Part 9: Prefer std::bit_width instead of std::countl_zero in js/src. r=sergesanspaille"

This reverts commit da9a43075a4336384741371cf33a38419751d645.

Revert "Bug 2020608 - Part 8: Replace CountTrailingZeroes with std::countr_zero. r=sergesanspaille"

This reverts commit d77ca273cb55bdde7b29dec5a29b5220f69f6043.

Revert "Bug 2020608 - Part 7: Replace CountLeadingZeroes32 with std::countl_zero or std::countl_one. r=sergesanspaille,nbp,profiler-reviewers,canaltinova"

This reverts commit 94929ae729fa299783219774a57114eb63e4057c.

Revert "Bug 2020608 - Part 6: Replace CountLeadingZeroes64 with std::countl_zero. r=sergesanspaille,profiler-reviewers,canaltinova"

This reverts commit 9b428168df32a9f37c24fc649d598dfa02b9377d.

Revert "Bug 2020608 - Part 5: Replace CountTrailingZeroes32 with std::countr_zero. r=sergesanspaille,manuel"

This reverts commit d3e77036b20a273669220d409b8e90ca9df7929e.

Revert "Bug 2020608 - Part 4: Replace CountTrailingZeroes64 with std::countr_zero or std::countr_one. r=sergesanspaille"

This reverts commit 6a7bcee10e93687f52ed409d2bead88d5842f4d4.

Revert "Bug 2020608 - Part 3: Replace CountPopulation32 with std::popcount. r=sergesanspaille"

This reverts commit e7b204177b59f0727f859922866a39b5707eb66e.

Revert "Bug 2020608 - Part 2: Replace CountPopulation64 with std::popcount. r=sergesanspaille"

This reverts commit 1d6f98a622fff5c98e9bd6fc56e7d747506709b9.

Revert "Bug 2020608 - Part 1: Add <bit> C++20 definitions to std.py. r=sergesanspaille"

This reverts commit 3ee4c6dcab8e628e9bef71d45aa33796064746e5.

UltraBlame original commit: 2a23ad974481c0e909d90bf22cefb4922fa6683a

+++ js/src/jit/loong64/Lowering-loong64.cpp
@@ -179,7 +179,7 @@ void LIRGeneratorLOONG64::lowerDivI(MDiv* div) {
     
     
     
-    int32_t shift = FloorLog2(uint32_t(rhs));
+    int32_t shift = FloorLog2(rhs);
     if (rhs > 0 && 1 << shift == rhs) {
       LDivPowTwoI* lir =
           new (alloc()) LDivPowTwoI(useRegister(div->lhs()), temp(), shift);
@@ -208,7 +208,7 @@ void LIRGeneratorLOONG64::lowerDivI64(MDiv* div) {
 void LIRGeneratorLOONG64::lowerModI(MMod* mod) {
   if (mod->rhs()->isConstant()) {
     int32_t rhs = mod->rhs()->toConstant()->toInt32();
-    int32_t shift = FloorLog2(uint32_t(rhs));
+    int32_t shift = FloorLog2(rhs);
     if (rhs > 0 && 1 << shift == rhs) {
       LModPowTwoI* lir =
           new (alloc()) LModPowTwoI(useRegister(mod->lhs()), shift);
