+++ js/src/jit/x86/Lowering-x86.cpp
@@ -204,7 +204,7 @@ void LIRGeneratorX86::lowerForMulInt64(LMulI64* ins, MMul* mir,
 
   if (rhs->isConstant()) {
     int64_t constant = rhs->toConstant()->toInt64();
-    int32_t shift = mozilla::FloorLog2(uint64_t(constant));
+    int32_t shift = mozilla::FloorLog2(constant);
     
     if (constant >= -1 && constant <= 2) {
       needsTemp = false;
